[Xamarin.Android.Build.Tasks] APK environment file does not contain XA_HTTP_CLIENT_HANDLER_TYPE in windows builds
Fixes https://bugzilla.xamarin.com/show_bug.cgi?id=44217

There is a slight oversight when we added the code to add the
default XA_HTTP_CLIENT_HANDLER_TYPE. If not environment files
were provided the default that was used did NOT include the
default XA_HTTP_CLIENT_HANDLER_TYPE.

In fact there were two code paths which dealt with defaults.
This commit removes the first code path which checks for a null
Environment array in favour of using the defaults which are added
at the end of the method. This means we maintain the defaults in
on place.