[Xamarin.Android.Build.Tasks] Add a Fast path for remving fixed directories. (#285)
Currently we loop through ALL the files in a directory and check
them all to see if they are writeable. What this commit does is
change it so we only do that IF we error the first time we try to
delete the directory.

If we can delete the directory first we just do. If not we make sure
everything is writable, and then of we fail we throw the error as usual.