[Mono.Android] Add a manual dispose fallback path when calling Dispose on a surfaced object fails (#288)
Previously we would simply call `Dispose` on each object surfaced to the runtime before shutdown so that no finalizer code could be run into a dead appdomain.

Unfortunately, because `Dispose` naturally calls `Dispose(bool)` before doing any low-level cleanup, user code can potentially throw exceptions which prevent the operation from happening correcly and (currently) triggers a process crash because of the unhandled exception.

This is the case when in the context of the Xamarin.Forms previewer where XF objects usually implement their own Dispose mechanism. Due to the "flat" nature of our cleanup loop, it may happen that we call `Dispose` on a XF object that in its own `Dispose(bool)` implementation also tries to invoke `Dispose` on an underlying object. The issue is that we may have already disposed it previously in our loop (since we have no clue on objects relationship with each other) causing an ObjectDisposedException to be thrown:

 Unhandled Exception:
 System.ObjectDisposedException: Cannot access a disposed object.
 Object name: 'Android.Widget.ListView'.
   at Java.Interop.JniPeerMembers.AssertSelf (Java.Interop.IJavaPeerable self) [0x00036] in /Users/builder/data/lanes/1196/cf38796b/source/Java.Interop/src/Java.Interop/Java.Interop/JniPeerMembers.cs:166
   at Java.Interop.JniPeerMembers+JniInstanceMethods.InvokeVirtualVoidMethod (System.String encodedMember, Java.Interop.IJavaPeerable self, Java.Interop.JniArgumentValue* parameters) [0x00002] in /Users/builder/data/lanes/1196/cf38796b/source/Java.Interop/src/Java.Interop/Java.Interop/JniPeerMembers.JniInstanceMethods_Invoke.cs:21
   at Android.Widget.AdapterView.set_OnItemClickListener (Android.Widget.AdapterView+IOnItemClickListener value) [0x00031] in /Users/builder/data/lanes/1196/cf38796b/source/monodroid/src/Mono.Android/platforms/android-24/src/generated/Android.Widget.AdapterView.cs:946
   at Xamarin.Forms.Platform.Android.ListViewAdapter.Dispose (System.Boolean disposing) [0x00050] in C:\BuildAgent3\work\82141f6bfad29a4c\Xamarin.Forms.Platform.Android\Renderers\ListViewAdapter.cs:332
   at Java.Lang.Object.Dispose () [0x00000] in /Users/builder/data/lanes/1196/cf38796b/source/xamarin-android/src/Mono.Android/Java.Lang/Object.cs:203
   at Android.Runtime.JNIEnv.Exit () [0x00051] in /Users/builder/data/lanes/1196/cf38796b/source/xamarin-android/src/Mono.Android/Android.Runtime/JNIEnv.cs:309

This commit adds a workaround for now that catches all exception on calling `Dispose` and log them. If an exception did happen, it then proceed onto trying to manually clean up the JLO instance by reproducing the second part of `Java.Lang.Object.Dispose` to clear the object's GREF and suppress its finalizer. This is far from "nice" but until a better integration with `Java.Interop.Runtime` can be devised it's the less intrusive method.