[mono-runtimes] Check for correct filename before `strip`ing
Fix a [Build][0] [Break][1]:

	Executing: ".../mxe/bin/x86_64-w64-mingw32.static-strip"  "../host-mxe-Win64/.dll"
	.../x86_64-w64-mingw32.static-strip: '.../host-mxe-Win64/.dll': No such file

This was a typeo introduced in commit a5b324d1 -- among others :-(:

	<Exec
	    Condition=" '$(Configuration)' != 'Debug' And '%(_MonoRuntime.DoBuild)' == 'true' And '%(_MonoRuntime.OutputMonoPosixHelperFilename)' != '' "
	    Command="&quot;%(_MonoRuntime.Strip)&quot; %(_MonoRuntime.StripFlags) &quot;$(OutputPath)\lib\xbuild\Xamarin\Android\lib\%(_MonoRuntime.Identity)\%(_MonoRuntime.OutputMonoBtlsFilename).%(_MonoRuntime.NativeLibraryExtension)&quot;"
	/>

One of these values is not like the other, and that's the problem.
`Condition` checks that
`%(_MonoRuntime.OutputMonoPosixHelperFilename)` isn't empty, but the
file we **strip** is `%(_MonoRuntime.OutputMonoBtlsFilename)`. When
building for `mxe-Win64`, `%(OutputMonoPosixHelperFilename)` is set,
but `%(OutputMonoBtlsFilename)` isn't, so we interpolate an empty
string, resulting in a filename which doesn't exist.

Fix the error by updating the `Condition` to check that
`%(OutputMonoBtlsFilename)` isn't empty.

[0]: https://jenkins.mono-project.com/view/Xamarin.Android/job/xamarin-android/140/
[1]: https://jenkins.mono-project.com/view/Xamarin.Android/job/xamarin-android/140/consoleText