[Xamarin.Android.Build.Tasks] Look in the AndroidSdk/extras folder for .aar files
The current Task always downloads data from google. This is regardless
of whether the files exist in the Android sdk directory already.
This commit changes the GetAdditionalResourcesFromAssemblies task
to look in the "extras" folder in the android sdk for the required
.aar file. This will mean if a user has an up to date sdk, there will
be almost no need to download the files from the internet.