Rewoked code in prvious commit to allow Google Play Services to use the AndroidSdk directory.
The prvious commit only added support for getting .aar files from the
`extras\android` directory. So if a user was using GPS it would still
download the files.

This commit reworks the system to make use of the known `extras` directories
when searching for files. This allows GPS to use this system as well.