[Xamarin.Android.Build.Tasks] Allow the cache'ing of the dex/jack compilation for support libraries.
Support libraries mostly contain classes.jar which needs to be converted to
a .dex format (or the .jack format). This can take some time. If we
pre convert the support libraries and cache the results we can cut down
the build time. dx.jar has the capability to merge in other .jar files
which contain classes.dex instead of .class files.