[Xamarin.Android.Build.Tasks] Look in the AndroidSdk/extras folder for .aar files (#296)
* [Xamarin.Android.Build.Tasks] Look in the AndroidSdk/extras folder for .aar files

The current Task always downloads data from google. This is regardless
of whether the files exist in the Android sdk directory already.
This commit changes the GetAdditionalResourcesFromAssemblies task
to look in the "extras" folder in the android sdk for the required
.aar file. This will mean if a user has an up to date sdk, there will
be almost no need to download the files from the internet.

* Rewoked code in prvious commit to allow Google Play Services to use the AndroidSdk directory.

The prvious commit only added support for getting .aar files from the
`extras\android` directory. So if a user was using GPS it would still
download the files.

This commit reworks the system to make use of the known `extras` directories
when searching for files. This allows GPS to use this system as well.

* Fixed the logging based on review feedback

* Fixed code format