[build] Don't use config.guess and use "$HOST_CC -dumpmachine" instead (#293)
The path to config.guess is dependent on distributions, so don't use
config.guess and use "$HOST_CC -dumpmachine" instead. It is also better
in point that it gives the result consistent with the compiler target. This
change is based on the following assumptions:

* $HOST_CC provides -dumpmachine option.
GCC and clang provide the option. Some dependencies such as LLVM expect
either will be used for compiling and it is likely to be safe to assume
that.

* $HOST_CC -dumpmachine and config.guess returns the same result.
It is likely because it won't build if the target architecture is
inconsistent.

* If the host architecture is 64-bit, it is compatible with i686.
It is likely to be safe to assume because `i386`, a command used before,
also assumes the same condition.