[tests] `make run-nunit-tests` needs to emit NUnit2 XML
Now that we can run NUnit tests on-device (48e3fc26), the next step is
to start running unit tests on Jenkins, which is best done through the
`make run-all-tests` target.

...or so I thought.

In actuality, that's not *quite* right, because `make run-nunit-tests`
will bail on the first failure, which is not at all desirable. We want
to run *all* the tests, not just some subset of tests that pass.

Fix the `RUN_NUNIT_TEST` define so that an "error exit code" is
"ignored." This allows all NUnit-based tests to actually execute.

Even better, it means that the `make run-nunit-tests` target won't
fail, which means `make run-all-tests` can actually progress to
running `make run-apk-tests`, which were being skipped because the
`Xamarin.Android.Build.Tests.dll` tests had some failures.

<heavy sigh>

That's only half the problem on Jenkins. The other half is that the
Jenkins **Publish xUnit test result report** step, using the
**NUnit-Version N/A (default)** pattern, was failing because the XSLT
used to process the NUnit XML output was emitting invalid XML:

	<?xml version="1.0" encoding="UTF-8"?>
	  /Users/builder/jenkins/workspace/xamarin-android-pr-builder/xamarin-android/packages/NUnit.ConsoleRunner.3.2.1/tools/nunit3-console.exe bin/TestDebug/Xamarin.Android.Build.Tests.dll --result=TestResult-Xamarin.Android.Build.Tests.xml -output=bin/TestDebug/TestOutput-Xamarin.Android.Build.Tests.txt
	  ...

That's certainly not valid XML...

The problem here was that `nunit3-console.exe` was emitting
NUnit **3** XML output, while the Jenkins XSLT only supports
NUnit **2** output. The fix here is twofold:

1. Update `pakcages.config` to also install the `NUnit.Console`
    and `NUnit.Extension.NUnitV2ResultWriter` packages.

    (Only the `NUnit.Console` package should be required, as it
    depends on the latter, but for some reason referencing *just*
    `NUnit.Console` didn't cause `NUnit.Extension.NUnitV2ResultWriter`
    to be installed on my local machine.)

2. Update the `nunit3-console.exe` invocation to specify the output
    format, which is a `;format=nunit2` *suffix* on `--result`:

        nunit3-console.exe "--result=TestResult.xml;format=nunit2" ...

(Using `;format=nunit2` *requires* the
`NUnit.Extension.NUnitV2ResultWriter` package.)

Next, when the `<CheckAdbTarget/>` ran the following command:

	$ adb shell getprop ro.build.version.sdk

When stdout contained an informational or error message:

	* daemon not running. starting it now on port 5037 *
	* daemon started successfully *
	error: device '(null)' not found

This was erroniously treated as "valid" output when
`CheckAdbTarget.SdkVersion` wasn't specified. Fix the
`<CheckAdbTarget/>` task so that informational and error messages are
ignored.

Finally, for currently unknown reasons, when the
`<RunInstrumentationTests/>` task is ran on Jenkins, the
`nunit2-results-path` may contain a `\r`, resulting in an error trying
to copy the file off the device:

	Tool .../adb execution started with arguments:   pull "/data/data/Mono.Android_Tests/files/.__override__/TestResults.xml^M" "../TestResult-Mono.Android_Tests.xml"
	...
	remote object '/data/data/Mono.Android_Tests/files/.__override__/TestResults.xml^M' does not exist

Fix this by using `string.Trim()` to remove whitespace characters from
`targetTestResultsPath`.

With all those changes in place, I *hope* Jenkins will be able to
(1) *run* all unit tests, and (2) *display* the results of those unit
tests in a nice, readable, graphical, form.