[mono-runtimes] Fix @(BundleItem) timestamps (#307)
The *intent* of the `_BuildUnlessCached` target within
`mono-runtimes.targets` is to (attempt to!) make building mono "free",
or at least, free *enough* so that other project files can have a
build-time dependency on `mono-runtimes.mdproj` so that build
artifacts can be (reasonably) consistent.

Unfortunately, that wasn't the case, with the net result being that:

	tools/scripts/xabuild /t:SignAndroidPackage src/Mono.Android/Test/*.csproj

would *constantly* attempt to re-build
`build-tools/mono-runtimes.mdproj`, adding (at least!) one minute as
we ran mono's `make`...to do nothing.

Why?

	Target _BuildUnlessCached needs to be built as input file '.../xamarin-android/external/mono/autogen.sh' is newer than output file '../../bin/Debug//include/armeabi-v7a/eglib/config.h'

The "problem" is that the `<Copy/>` task copies over the timestamp as
well, so even though we're (possibly) creating a new file, the
timestamp on the created file will match the source file.

Normally this is fine. In this case, we're dealing with `eglib`, which
is a "nested" ("delegated"?) `configure` invocation from mono's
`configure`. It *appears* that if (when) mono is updated, "normal"
mono subdirectories will "properly" honor `--enable-maintainer-mode`
and re-run mono's `configure`, which will in turn update the timstamp
on any `configure`-time generated files.

This doesn't appear to be the case for `mono/eglib`; when `mono` is
updated, `mono/eglib` does *not* re-run `configure`, and thus no
timestamps are updated to follow suit.

Fix this by *explicitly* `<Touch/>`ing all copied files so that we
ensure they have a "recent" timestamp ("now"), which should ensure
that the file's timestamp is newer than that of e.g. `autogen.sh`.

Additionally, there was a typo: Commit a5b324d1 used a
`@(_InstallRuntimesOutput)` item group, but there is no such item
group; it's actually `@(_InstallRuntimeOutput)` (no `s`).
Fixity fix this.

Finally, commit a5b324d1 forgot to add `libmono-btls-shared` to
`@(BundleItem)`, preventing it from being added to the bundle.
Fix this oversight.