[monodroid, tests] Export CreateZStream(); Fix tests
PR build [#297][0] reported 8 failures. Fix two of them:

* `GzipStreamTest.Compression`
* `SslTest.SslWithinTasksShouldWork`

`GzipStreamTest.Compression` was failing because the symbol
`CreateZStream` couldn't be found:

	System.EntryPointNotFoundException : CreateZStream
	  at (wrapper managed-to-native) System.IO.Compression.DeflateStreamNative:CreateZStream (System.IO.Compression.CompressionMode,bool,System.IO.Compression.DeflateStreamNative/UnmanagedReadOrWrite,intptr
	  ...

`libmono-android*` is compiled with `-fvisibility=hidden` so that all
symbols are not exported from the native library by default. In order
for a symbol to be exported, the symbol needs to be annotated with:

	__attribute__ ((visibility ("default")))

`CreateZStream()` *looks* like it should be annotated like this:

	/* external/mono/support/zlib-helper.c */
	MONO_API ZStream *CreateZStream (gint compress, guchar gzip, read_write_func func, void *gchandle);

`MONO_API` is (usually) a macro that expands to
`__attribute__((visibility("default")))`, so one would think that this
would work.

The problem is that it *doesn't* work, and the reason is because no
header that `zlib-helper.c` includes provides a definition for
`MONO_API`, so it is insted defined as *nothing*. Consequently,
`CreateZStream` isn't exported, and code that requires it breaks.

The fix is twofold:

1. Update `$(LOCAL_C_INCLUDES)` within `src/monodroid/jni/Android.mk`
    so that `bin/$(Configuration)/include` is searched for header
    *first*.

2. Update the `_GenerateIncludeFiles` target within
    `src/monodroid/monodroid.targets` to copy a custom `config.h` into
    `bin/$(Configuration)/include`.

The custom `config.h` -- `src/monodroid/jni/config.h` -- ensures that
`MONO_API` has the correct value when `<config.h>` is included, which
is a header that `zlib-helper.c` includes, thus ensuring that
`CreateZStream()` is exported.

"While we're at it", update `Android.mk` so that instead of
"hard-coding" the relative path to the mono checkout, we instead use
`$(MONO_PATH)`, set to `$(MonoSourceFullPath)`.
(See e.g. d205cab2.)

The problem with the `SslTest.SslWithinTasksShouldWork` test was that
it was using a no longer maintained URL, `pipeline.internalx.com`.
Fix the test to instead use `tlstest-1.xamdev.com`.

Finally, update `HttpClientIntegrationTests.cs` to exclude the same
tests that we exclude internally, as many of those tests are "flaky"
or "unreliable": they'll pass sometimes, and fail others, and the
inconsistency is maddening.

[0]: https://jenkins.mono-project.com/view/Xamarin.Android/job/xamarin-android-pr-builder/297/testReport/