[Xamarin.Android.Build.Tasks] Rework the download cache to share files.
One of the problems we have with the current system is the same
.aar files are extracted into multiple locations. These locations
are based on the Assembly name. e.g

	Xamarin.Andriod.Support.v7/

This just wastes disk space.

This commit reworks the system to detect if its a "xamarin" assembly
and to extract the data to an "extras" directory instead. The
files are still in the versioned folder so there is no danger of
overwriting existing files.

Eventually this will be replaced with the ability to reference the
.aar files directly from the android project in which case they will
be extract into the local resourcecache directory.