[Xamarin.Android.Build.Tests] Fix the Tls tests (#310)
This commit fixes the Tls tests. The first problem was that
the Ionic.Zip dll was not being copied local. So Ionic has
been replaced with a LibZipSharp implementation.

The second was the path to the apk needed to use the
Root property. This property provides the root path to the
unit test working directory.

Finally, the code to resolve the XA_BUILD_OUTPUT_PATH was
incorrect. As a result the unit tests were using the installed
mono/xamarin-android not the built one.