[Xamarin.Android.Build.Tasks] Rework target order so _ValidateAndroidPackageProperties is called early.
Context https://bugzilla.xamarin.com/show_bug.cgi?id=46931

Rather than have a bunch of DependsOn calls to _ValidateAndroidPackageProperties we
just need to rework the order so that this target is called early as part of the
Build target.

Once we do this we can revert the commit 5fc80f27 which broke the build.