[msbuild][export-api-merged-xml] implement SourceJar ordering.
Here is the basic design idea behind this (first step) implementation:

----

API set auto detection is supported when the parent directory of each
package is a valid maven repository, which is structured as:

	.../m2repository/{groupID}/{artifactID}/{version}/{file.aar}

This should be the normative structure. Libraries with different
directory structure can be transformed to become equivalent to this
structure by making use of file links.