[Xamarin.Android.Build.Tasks] Fix <Aot/> to work on background threads. (#329)
The recent commit 7d8bb92b reworked the Aot task to run the cross
compilers in parallel. However this commit did NOT take into account
that the MSBuild Logger does not like being used from other threads.

So from the command line we end up with exceptions being raised (occasionally)
from the Logger. I suspect within VS that patch would have caused a
complete IDE lockup. This is because VS does not like it when you
Log messages/errors from a thread other than the UI thread.

This commit reworks the Aot Task to derive from the AsyncTask. This
task implements a number of helper methods which resolve the
Logger/UI Thread issues.