[msbuild] Add new AndroidSkipJavacVersionCheck property.
We check javac version and adjust source compatibility to 1.7 if the
version is 1.8, and proguard or multidex is used (because, they are
incompatible with Java8).

This, being the mandatory behavior is, however, problematic.

When we supply custom proguard as a NuGet package to resolve the Java8
incompatibility issue, we still expect Java8 is used (otherwise,
support library applications do not compile). But the above prevents that.

So, skip javac version check when AndroidSkipJavacVersionCheck property
is 'True'.

My facebook proguard NuGet package already has this property as true
in its MSBuild props file. So you can add this in your Android 7.0+
project and enable proguard or multidex, once you get this change merged.
https://www.nuget.org/packages/name.atsushieno.proguard.facebook/5.3.2.2