[Xamarin.Android.Build.Tasks] Archiving application is archiving old builds (#333)
Fixes https://bugzilla.xamarin.com/show_bug.cgi?id=48508

The problem here was the output from the BuildApk task was not
being included as an input for the task which copied the apk to
the output directory. So when it was being updated (i.e not created)
it was not being flagged as changed by the build system.