[tools/scripts] Add mono-symbolicate to the build system.
We were missing mono-symbolicate as a result release builds would
often fail when it tries to symbolicate the assemblies.

This commit adds a new Target to build the mono-symbolicate.exe from
the mono tree and place it in

	bin\$(Configuration)\lib\mandroid

it also adds two scripts, one for unix based systems and one for windows.
These scripts will end up in

	bin\$(Configuration)\bin

and will be run by the build tasks from that location.