Add support for redirect to url without protocol
This commit adds support for redirecting to urls that
do not specify a scheme. In such cases, the scheme
of the previous request is used.

For example, a request to http://example.com/redirect
may return a 301 redirect to //example.com/new_site.
In this case, we should keep speaking http and retrieve
http://example.com/new_site

Relative urls without protocol are supported according to RFC 3986,
so this commit adds a comment with a link to the RFC.