[Xamarin.Android.Build.Tasks] Add a new AndroidPrepareForBuild Target for unit tests
Some of our unit tests dont need to build but do need to resolve
the android SDK and other properties. We did us the _ValidateAndroidPackageProperties
target to make sure all of the required properties were in place.

However commit 8337d21e altered the build process so that target is
run earlier in the build process. As a result the required properties
are not initialised.

This commit adds a new "AndroidPrepareForBuild" target which ensures that
the correct targets are run for tests.