[Xamarin.Android.Build.Tasks] Asyncify Aapt Task
We currently use MSBuild Batching to call the aapt tool
when processing the resources. This results in about 12 seperate
calls to aapt when building a standard Xamarin.Forms app.

The problem with Batching is that it is sequential. MSBuild and
xbuild are unable to run those batches in parallel. So this
commit reworks the Aapt task to be an AsyncTask. Which can take
multiple AndroidManifest files at once. It will then spin up
aapt in parallel rather than running them sequentially.

The difference between the two systems is clear

Using Batching:

6156.275 ms  Aapt  10 calls

Using Async:

3207.201 ms  Aapt  3 calls

This is a first build on a vanilla Xamarin.Form app with no changes.