[Android] Fix friendly name construction for extra appdomain.
In 17bdeb38cf0d2 this line was changed to use a pre-cached version of an API level.

This doesn't have any sense in that case though because the friendly name is
specifically constructed to identify the Android API level for which it's
being created. With this change the value is not actually representing this
anymore but instead is whatever API level value that was cached previously.

This commit thus reverts the line to use the previous method which fetch the info
from the passed ClassLoader instance ensuring the name is always valid.