[Xamarin.Android.Build.Tasks] Don't deploy Facade assemblies (#337)
A change "somewhere" has recently causes this issue. The problem is
in Release mode we now include the Facade assemblies in the package.
Which is something we definately should NOT be doing.

That said we do need the Facades in order for Linking and Aot to
work since Cecil needs to be able to load them.
So what this commit does is make sure the `ResolveAssemblies` task
adds metadata for the Facades. We can then use that in the `BuildApk`
Task to filter out the assemblies we do not want to ship.