Query Android for DNS server addresses
Mono parses `/etc/resolv.conf` on Unix systems in order to retrieve the list of
DNS server IP addresses and return it from the
`IPInterfaceProperties.DnsServers` property.
On Android, however, `resolv.conf` doesn't exist and in consequence the DNS
server list is never populated. Instead, on Android we need to read the IPs from
the system properties named `net.dnsX` where X is a number between 1 and 8,
inclusive.
This code implements the XA side of the fix.

Part 1 of fix for https://bugzilla.xamarin.com/show_bug.cgi?id=48016